Remove 'node_class' argument in caching config functions
Change the 'identifier' to always refer to the process_type. Use
the 'fnmatch' stdlib to allow globbing in the configuration.

This is technically a breaking change, because the interface
of 'get_use_cache', 'enable_caching', 'disable_caching' no longer
allows the 'node_class' interface. However, the previous
version silently did nothing when this argument is passed, so
this just makes the issue more apparent.