`BaseRestartWorkChain`: require process handlers to be instance methods (#3782)
The original implementation of the `register_process_handler` allowed an
unbound method to be bound to a subclass of `BaseRestartWorkChain`
outside of its scope. Since this also makes it possible to attach these
process handlers from outside of the module of the work chain, this will
run the risk of the loss of provenance. Here we rename the decorator to
`process_handler` and it can only be applied to instance methods. This
forces the handlers to be defined in the same module as the work chain
to which they apply.