Implement the logic for getting the most specific identifier
When using globbing in the caching configuration, we want the
ability to override a general glob (e.g. aiida.calculations:*)
with a more specific one (e.g. aiida.calculations:aiida-diff.*).
This commit implements the logic and tests for this behavior.
To check which identifier pattern is more specific, the patterns
are fnmatch'ed against each other. If all other patterns match
a given pattern, it is the most specific.