Fix bugs in `Node._store_from_cache` and `Node.repository.erase` (#3777)
The `_store_from_cache` needed to erase the content of its sandbox
folder before copying over the content of the cache source. Currently,
the existing contents are mixed with with the contents to be copied in.

In fixing this, we discovered an issue in `Node.repository.erase`, where
an unstored node would try erasing the (non-existent) folder in the
permanent repository, instead of the `SandboxFolder`.