Avoid deadlocks when retrieving stdout/stderr via SSH
Fixes #1525

The simple shift of the line `retval = channel.recv_exit_status()`
below `stdout.read()` and `stderr.read()`, while partially improving
the situation, is not enough (see discussion in #1525).
This instead reads in chunks both stdout and stderr, alternating.
Tests show that this does not hang anymore for large files.