verdi computer/code duplicate now reuses prepend_text and append_text
This required adding a new ContextualDefaultOption class
to get a default callable that also receives the context.
Moreover, a but was fixed in the function that prompts
for the prepend/append_text, as if the file was not touched/modified,
instead of reusing the text provided in input, an empty text was used instead.