Explicitly match only '*' wildcards
The 'fnmatch' implementation allows using '?' (any single
character) or '[abc]' (any of the characters in brackets) matches.
These are problematic when trying to determine the 'most specific'
match to a given identifier. To avoid this, we instead implement
matching of _only_ '*' wildcards with 're'. This is similar to how
'fnmatch' is implemented, except for removing the extra logic
needed for '?' and '[]' matches.