Write migrated config to disk in `Config.from_file`
When an existing configuration file with an outdated schema was loaded
from disk through `Config.from_file`, as happens in the initialization
call of `aiida.manage.configuration.load_config`, the content was
properly migrated in memory but the changes were not written to file.
This caused the migration to be performed each time.