Explicitly call _store_from_cache in test
Because the _cachable attribute is now set to False for Data
nodes, the test for the '_store_from_cache' method failed because
caching was not used. Instead of creating a valid cachable node
(which is slightly complicated), we just call _store_from_cache
directly to circumvent the 'is_valid_cache' and '_cachable'
checks.