Disable caching for the `Data` node subclass (#3807)
Caching was enabled for the `Data` node sub class, even though with the
definition of "caching" in AiiDA that doesn't really make sense. It is a
manner to prevent having to reexecute a calculation that already has
been executed before, which clearly does not apply to `Data` nodes.
Since it was simply a no-op, it was left in. However, it turns out it
can actually create problems when a cloned data node is stored that has
changed some mutable attributes. Since those are not included in the
hash, the hash of the clone is the same as that of its source and the
the caching is activated replacing the clone with the source, undoing
the changes to the mutable attribute. All of this is solved by simply
disabling caching for `Data` nodes.