Caching: fix configuration spec and validation (#3785)
Remove the `node_class` argument in caching config functions and change
the `identifier` to always refer to the `process_type` of a node.

The identifiers in the caching configuration now have to be either valid
entry point names or the process type of a node, which is the join of a
python module path and the resource, like function or class. The
identifiers are now also allowed to contain the wildcard character `*`.
This allows globbing to match groups of entry point names.

This is technically a breaking change, because the interface of
`get_use_cache`, `enable_caching`, `disable_caching` no longer allows
the `node_class` interface. However, the previous version silently did
nothing when this argument is passed, so this just makes the issue more
apparent.