`ExitCode`: make the exit message parameterizable through templates
Often one would like to define an exit code for a process where the
general gist of the message is known, however, the exact form might be
slightly situation dependent. Creating individual exit codes for each
slight variation is cumbersome so we introduce a way to parameterize the
exit message. To do this the `ExitCode` had to be changed from named
tuple to a proper class, such that the `__call__` method could be
implemented. This allows a mechanism where calling an exit code with a
parameterized exit message can be concretized. The keyword arguments are
used for the string interpolation and, importantly, a new exit code
instance is returned. Example:

    exit_code_template = ExitCode(450, 'the parameter {parameter} is invalid.')
    exit_code_concrete = exit_code_template(parameter='some_specific_key')

The `exit_code_concrete` is now unique to `exit_code_template` except
for the message whose parameters have been replaced.