Restore `ExitCode` to being a namedtuple
Changing the `ExitCode` from a namedtuple to a class breaks backwards
compatibility since the class won't have the functionalities of a tuple
such as expanding an instance in its attributes. By subclassing from the
namedtuple definition we keep all that functionality while still being
able to implement additional methods, such as in this case `format`.

Co-Authored-By: Dominik Gresch <greschd@users.noreply.github.com>