`ExitCode`: make the exit message parameterizable through templates (#3824)
Often one would like to define an exit code for a process where the
general gist of the message is known, however, the exact form might be
slightly situation dependent. Creating individual exit codes for each
slight variation is cumbersome so we introduce a way to parameterize the
exit message. We implement a `format` method that will format the message
string with the provided keyword arguments and returns, importantly, a
a new exit code instance with the new formatted message.

    exit_code_template = ExitCode.format(450, '{parameter} is invalid.')
    exit_code_concrete = exit_code_template(parameter='some_specific_key')

The `exit_code_concrete` is now unique to `exit_code_template` except
for the message whose parameters have been replaced.

To enable to implement a new method, we had to turn the `ExitCode` from
a named tuple into a class, but to keep backwards compatibility, we sub
class from a namedtuple, which guarantees that the new class behaves
exactly like a tuple.