Ensure log messages are not duplicated in daemon log file
The logging for the daemon was configured with a handler both for
writing to stdout `console` as well as a rotating file handler that
writes directly to the daemon log file. The last one is the intended one
but since the default `console` handler was also kept, its output also
ended up in the daemon log file, duplicating every message.