Fix the `RotatingFileHandler` configuration of the daemon logger
The logging configuration for the daemon did already include the
definition of a `RotatingFileHandler`, however, the logs were not
actually being rolled over when the maximum size was hit. The problem
was because the argument `backupCount` was not defined. As the python
documentation of the `logging` module states:

  but if either of `maxBytes` or `backupCount` is zero, rollover never
  occurs, so you generally want to set `backupCount` to at least 1,
  and have a non-zero `maxBytes`.

Setting `backupCount` to 10, fixes the problem. The size of each file is
set to 10 MB. This should then keep at most 100 MB of log files for each
profile.