Implement functionality to immigrate already completed `CalcJobs`
When people start using AiiDA they typically already have many
calculation jobs completed without the use of AiiDA and they wish to
import these somehow, such that they can be included in the provenance
graph along with the future calculations they will run through AiiDA.

This concept was originally implemented for the `PwCalculation` in the
`aiida-quantumespresso` plugin and worked, but the approach required a
separate `CalcJob` implementation for each existing `CalcJob` class that
one might want to import.

Here we implement a generic mechanism directly in `aiida-core` that will
allow any `CalcJob` implementation to "immigrate" already completed
jobs. The calculation job is launched just as one would launch a
normal one through AiiDA, except one additional input is passed: a
`RemoteData` instance under the name `remote_folder` that contains the
output files of the completed calculation. The naming is chosen on
purpose to be the same as the `RemoteData` that is normally created by
the engine during a normal calculation job run.

When the engine detects this input, instead of going through the normal
sequence of transport tasks, it simply performs the presubmit and then
goes straight to the "retrieve" step. Here the engine will retrieve the
files from the provided `RemoteData` as if they had just been produced
during an actual run. In this way, the process is executed almost
exactly in the same way as a normal run, except the job itself is not
actually executed.