additional validation for cli arguments
 * Start validating
   * hostnames
   * emails
   * entry point names

 * InteractiveOption: validate empty reply

The `InteractiveOption` of the CLI always allowed users to pass `!` in
order to avoid setting a value, bypassing any further validation that
may be provided by the `type`.

Change this behavior such that entering `!` results in an empty string
being passed further to the `type`. Only if conversion by the type
passes as well should a `None` value be returned.

 * increase code reuse in options
 * move IdentifierParamType back to how it was before
 * add test for verdi code duplicate
   return empty string, if passing `!` into string-type parameters.