Move `aiida.manage.external.pgsu` to external package `pgsu` (#3892)
One of the main issues with `verdi quicksetup` is that it is non-trivial
to figure out how to connect as the PostgreSQL superuser in a wide
variety of operating systems and PostgreSQL setups. This PR factors out
the code which takes care of this connection into a separate package, 
called `pgsu` which can then be tested on a wide variety of setups using
continuous integration.