Remove superfluous `ERROR_NO_RETRIEVED_FOLDER` from `CalcJob` subclasses
The `ERROR_NO_RETRIEVED_FOLDER` is now defined on the `CalcJob`
base class and the `CalcJob.parse` method already checks for the
presence of the retrieved folder and return the exit code if it is
missing. This allows us to remove the similar exit codes that are
currently defined on the calculation plugins shipped with `aiida-core`
`ArithmeticAddCalculation` and `TemplateReplacerCalculation` as well as
the check for the presence of the `retrieved` output from the
corresponding parsers. The fact that is now checked in the `CalcJob`
base class means that `Parser` implementations can assume safely that
the retrieved output node exists.