Add export archive migration for `Group` type strings
The `Group` entity class is now subclassable, for which the type string
of existing entries in the database had to be changed through a
migration. Here we add the corresponding migration for existing export
archives. The only required change is to map the type string of groups.

To test this migration we use the existing export archives in the module
`tests/fixtures/export/migrate`. They did not contain any group entities
so `export_v0.1_simple.aiida` was updated first and had four groups
added, one for each of the migrated type strings. This initial archive
was then migrated to each subsequent version, step by step, using the
command `verdi export migrate --version`.

Also fixed and migrated `tests/fixtures/graphs/graph1.aiida` which was
corrupt and could not be migrated automatically, because it contained a
process node with an active process state.