Add support for subclassing in `Groups` to the `QueryBuilder` (#3903)
This now allows to search for instances of `Group` and all its
subclasses through the `QueryBuilder` just as implemented for the `Node`
class. The subclassing is based on the `type_string` of the group
instances, which in turn is defined by the entry point name. When
querying for `Group` instances through the `QueryBuilder`:

    QueryBuilder().append(Group).all()

by default subclassing is enabled and should return all instances of a
subclass of `Group` as well. The same goes for subclasses of `Group`
themselves. Imagine a plugin defines the following entry points:

    'plugin.sub = aiida_plugin.groups:SubGroup'
    'plugin.sub.a = aiida_plugin.groups:SubAGroup'
    'plugin.sub.b = aiida_plugin.groups:SubBGroup'

The following query:

    QueryBuilder().append(SubGroup).all()

will return all instances of `SubGroup` but also of `SubAGroup` and
`SubBGroup`. This is because their entry point names both start with the
entry point name `plugin.sub` of the class that is being appended.

Note that in order for all group instances to be returned when querying
for the `Group` base class, a small hack had to be applied in the query
builder code building the filters. Note that the entry point name of
`Group` is `core` and so with the rules described above, the plugin
subclasses would never be matched since their entry point names do not
start with `core`. This is a problem and could be fixed by making the
type string of base group instances an empty string, however, this would
require using an empty string for the `Group` base class which is not
allowed in Python. To still provide the desired behavior, when querying
for `Group` instances the `type_string` filter is set to match anything.