Deduplicate code for tests of archive migration code
Each archive migration was being tested on a archives included in this
repo in `tools/fixtures/export/migrate` and the test was always the
same. Simply take an archive, migrate it to the next version with the
appropriate method and check that both metadata and data dictionaries of
the migrated data match that of the reference archive. This same test
was implemented for each migration method, but has now been centralized
in a single test using pytest to parametrize the versions.

The testing on external archives from `aiida-export-migration-testing`
has also been streamlined a bit, by defining a new base class called
`ArchiveMigrationTest` that provides some utility methods.