Deprecate `--group-type` option in favor of `--type-string`
The `--group-type` option in `verdi group list` and `verdi group path ls`
has been deprecated and superseded with `-T/--type-string`. The recently
added support for subclassing of the `Group` ORM class has made it
interesting to be able to query for specific subclasses based on the
type strings of groups. Since these can be hierarchical it is useful to
support the inclusion of SQL wildcards `%` and `_` to perform like instead
of exact matching.

Note that this new functionality of filtering is only available for
`verdi group list` and not for `verdi group path ls`, as the `GroupPath`
utility does not support subclassing yet.