Add support for subclassing to `GroupParamType` CLI param type
This will allow options and arguments that are supposed to match `Group`
instance, to narrow the scope of subclasses that are to be matched.
For example, an option or argument with the following type:

  GroupParamType(sub_classes=('aiida.groups:core.auto',))

will only match group instances that are a subclass of `AutoGroup`. Any
other `Group` subclasses will not be matched.