Remove the equality operator of `ExitCode` (#3940)
The operator implementation assumed that `other` would always be an
instance of `ExitCode` as well and so therefore was guaranteed to
contain the same attributes. However, comparing it to a bare tuple would
also invoke the `ExitCode.__eq__` method and would raise and
`AttributeError` as a result. The desired behavior for the equality
operator is that it returns True for any other tuple of the same length
and content. Since this is exactly what is implemented by the tuple
base type, we do not have to override it at all.