`CalcJob`: make submit script filename configurable through options (#3948)
The name of the job submission script was hard coded to `_aiidasubmit.sh`
which not only makes it less flexible but it also required the value to
be passed in a long series of calls from `CalcJob.presubmit` all the way
through the transport tasks to the `submit_calculation` method in the
`aiida.engine.daemon.execmanager` module.

This value is now the input `metadata.options.submit_script_filename`
with the default being the same. This allows to remove the value from
the return value of `CalcJob.presubmit` and the call-chain through all
the transport tasks, because `submit_calculation` can now simply
retrieve the correct value directly from the `CalcJobNode`.

This change also fixes a bug in the `task_upload_job` that returned the
wrong value if the idempotency check was hit to see if the method had
already been executed. This can happen if the daemon is stopped before
the next transport task is reached but after the upload has already been
executed. This check prevents it from being executed again by simply
returning. However, it was supposed to return the `calc_info` and
`script_filename` normally returned by the `presubmit` call, but it
would no longer have these variables and so returned `True`. The caller
`Waiting.execute` would expect a tuple and except.