Close SQLA session after every REST API request
This is needed due to the server running in threaded mode, i.e.,
creating a new thread for each incoming request. This concept is great
for handling many requests, but crashes when used together with AiiDA's
global singleton SQLA session used, no matter the backend of the profile
by the `QueryBuilder`.

Specifically, this leads to issues with the SQLA QueuePool, since the
connections are not properly released when a thread is closed. This
leads to unintended QueuePool overflow.

This fix wraps all HTTP method requests and makes sure to close the
current thread's SQLA session after the request as been completely
handled.

Use Flask-RESTful's integrated `Resource` attribute `method_decorators`
to apply `close_session` wrapper to all and any HTTP request that may
be requested of AiiDA's `BaseResource` (and its sub-classes).

Additionally, remove the `__init__` function overwritten in
`Node(BaseResource)`, since it is redundant, and the attributes `tclass`
is not relevant with v4 (AiiDA v1.0.0 and above), but was never removed.
It should have been removed when moving to v4 in 4ff2829.

Concerning the added tests: the timeout needs to be set for Python 3.5
in order to stop the http socket and properly raise (and escape out of
an infinite loop). The `capfd` fixture must be used, otherwise the
exception cannot be properly captured.

The tests were simplified into the pytest scheme with ideas from
@sphuber and @greschd.