Fix bug in `verdi daemon restart --reset`
The command `verdi daemon restart --reset` was excepting when it was
invoking the `verdi daemon start` command. The reason being that recently
the `number` argument of `verdi daemon start` was changed to not define
a default directly, but to go through a callback. This would then set
the default based on a configuration option for the current profile.

However, due to a bug in `click`, parameter callbacks are not called
when invoking a command through `Context.invoke`, which caused the
`number` argument to be `None` which triggered an exception. Until that
bug is fixed in `click`, which has issue #950, we add a workaround to
manually defining the default value, that would normally have been done
by the argument callback, and explicitly passing it in the invoke call.