Fix exception handling in commands calling `list_repository_contents` (#3968)
The commands `verdi node repo ls` and `verdi calcjob inputls/outputls`
were passing the caught exception object straight into the echo call
without casting it to a string first, which would incur another
exception by itself. On top of that, they were catching `ValueError`
exceptions, which, as far as I can tell, would never be thrown. Instead
if the path does not exist a `FileNotFoundError` will be thrown. The
commands now catch this instead and print a useful message. This failure
path of the commands was not tested, so the bug went unnoticed.

The problem of raw exception objects being passed to `echo_critical`
were present in some other commands as well, which have now been fixed.