Docs: flesh out skeleton of "How-to Guides"
This section is subdivided in 5 main categories:

 * How to run external codes
 * How to run multi-step workflows
 * How to work with data
 * How to manage your installation
 * How to package plugins

The naming has been purposefully kept free of AiiDA specific jargon as
these guides need to be quickly findable by relatively new users that
after having run through a tutorial want to quickly know how to
accomplish certain specific goals.

Each section has its own document that will be presented as its own
page with a number of subsections that present guides for specific
goals. The sections and subsections are all created with the
corresponding link labels for cross-referencing.