Fix autocompletion for LinkManager and AttributeManager
In the managers, exceptions were not caught properly.
For instance, `getattr(calc.inputs, 'value')` was not return
an AttributeError if `'value'` does not exist, and as a
consequence `getattr(calc.inputs, 'value', None)` was raising
instead of returning `None`.

Similarly, I fixed `calc.inputs['value']` to raise a KeyError
for not-existing `value`.

This is now addressed by defining a new compound exception,
that inherits both from NotExistent (for backwards-compatible reasons)
and from the correct base exception of python (AttributeError or
KeyError).

The AttributeManager was not having Tab completion for a different
reason, `__dir__` was returning tuples of dict items, rather than
just the keys.

Now these are fixed and tests are added (I cannot really test
the TAB-completion functionality, but at least I test that the
values of `dir()` and the exceptions raised are the correct ones).

Fixes #3984