Docs: move in section "Topics - Processes"
The content that was envisioned for this section was originally divided
over the "Concepts" and "Working" sections, which were itself roughly
divided in information on "Processes", "Calculations" and "Workflows".
The original idea was to nest the material on calculations and workflows
within the global processes topic, since in the code the former are also
sub classes of the latter. However, since there are also documentation
on processes in general, I was not able to make this hierarchy work.

Instead, "Calculations" and "Workflows" are now individul topics on the
same level as "Processes". Each of these topics are roughly subdivided
in a "Concepts" and "Usage" section, where the content from the latter
came from the original "Working" directory. The concepts explains the
concepts of the topic, where the usage really exposes and explains the
API with all the nitty gritty details.