Fix timeout in `tests.cmdline.commands.test_process:test_pause_play_kill` (#4052)
The test `tests.cmdline.commands.test_process:test_pause_play_kill`
was failing in local setups with a time out. The problem was that the
daemon, launched as a subprocess in the `setUp`, would fail to load the
submitted process, because the class could not be imported. This is
because it is defined in the `tests.utils` module, which would not be
part of the Python path of the subprocess.

The `tests` module is automatically added to the path by `pytest`,
however, it was not being forwarded to the subprocess. Passing the
current environment through the `env` argument solves this. The only
additional change is to include the `PYTHONPATH` variable in the env
copy created by `aiida.cmdline.utils.common.get_env_with_venv_bin`.
These two changes guarantee that the test process classes are importable
by the daemon that is running in the subprocess.