Fix tab-completion for `LinkManager` and `AttributeManager` (#3985)
In the managers, exceptions were not caught properly. For instance,
`getattr(calc.inputs, value)` was not return an AttributeError if `value`
does not exist, and as a consequence `getattr(calc.inputs, 'value', None)`
was raising instead of returning `None`.

Similarly, `calc.inputs['value']` is fixed to raise a `KeyError` for
not-existing `value`.

This is now addressed by defining a new compound exception, that
inherits both from NotExistent (for backwards-compatible reasons) and
from the correct base exception of Python (AttributeError or KeyError).

The `AttributeManager` did not have Tab completion for a different
reason, as `__dir__` was returning tuples of dict items, rather than
just the keys.

Now these are fixed and tests are added TAB-completion functionality
cannot really be tested but at least it is tested that the values of
`dir()` and the exceptions raised are the correct ones.