Add a progress bar for export and import related functionality (#3599)
When provenance graphs become large, exporting and importing the
resulting archives can become time consuming operations. To give the
user more feedback of how the operation is progressing, a progress bar
is added. The implementation is provided by the `tqdm` library which is
therefore added as a new dependency.

Since the export and import code is still not written in a modular way,
to allow the progress bar to get access to the inner parts and provide
information that is granular enough, it is implemented as a global
singleton that the export and import functions fetch. This is to prevent
having to pass the progress bar instance around in method calls. This is
still not an ideal solution and in the future, this should be replaced
with hooks that methods can call in order to update their status.

A similar problem is faced with the logging of textual progress in the
export and import functions that are done haphazardly with print
statements, even though it concerns module functions and not CLI facing
code. This has been changed to go through logging instead. The log level
is temporarily updated based on the `silent` argument that the export
and import functions take.

Finally, the function signature of `export` and `export_tree` have been
changed. The `what` and `outfile` arguments have been deprecated and
replaced by `entities` and `filename`, respectively. A deprecation
message is printed if they are used and they should be removed with the
release of `aiida-core==2.0.0`.