Add highlight_classes argument in graph visualization
the argument highlight_classes is added in method
`recurse_ancestor` and `recurse_descendants` in
Provenance Graph visualization for the purpose of
 quickly spotting the desire node.

The target classes to be highlight are passed to the
argument as a class name string, it was compared with
the label of node. So the node label getting function
is isolated to be reused.