Increase the minimal required version of setuptools to 42.0.0 [py35] (#4085)
This is necessary in order for the requirements of Python to be properly
enforced by `setuptools`. A common problem that occurs when this is not
respected is where a new `numpy` pre-release is released and the setup
of `pymatgen` installs it and when not compatible with the Python version
the build fails. The newer version of setuptools should respect the
requirements that we impose.