Add support for binary files to the various `verdi` cat commands (#4077)
The commands `verdi calcjob inputcat`, `verdi calcjob outputcat` and
`verdi node repo cat` were implemented such to only be able to deal with
non-binary files. They are changed to simply copy the content of the
target file directly to the stdout buffer. This fixes two problems: the
command no longer assumes anything about the file content and just
copies the bytes as is, and the copy is performed in chunks preventing
from the whole content being loaded in memory.