Fix bug in `Code.get_full_text_info` (#4083)
This method was not tested and so it went unnoticed that it was still
using the `_get_folder_pathsubfolder` repository method that has been
removed in `aiida-core==1.0.0`. Direct access to the folder underlying
the node repository is no longer allowed and instead the API should be
used to inspect the available objects, in this case `list_objects`.