Increase the efficiency of the `SqlaGroup.nodes` iterator (#4094)
The iterator was calling dbnodes.__iter__() which was violting
the lazy loading of the nodes as they are needed during iteration
leading to a huge load time for big groups even when only a slice
is demanded. For reference, on a group with roughly half a million
nodes, getting a single one through group.nodes[0] went from
taking a minute to a few milliseconds.