Make the RabbitMQ connection parameters configurable
Up till now, the URL to connect to the RabbitMQ server was hardcoded.
This means it could only connect to the localhost over the standard port
and with the default credentials. Certain users, will require to deploy
RabbitMQ on a different machine than the AiiDA instance so the server
details should be configurable. Since this will no longer guarantee that
the RabbitMQ server is running on localhost, it should also be possible
to use SSL by changing the protocol from `amqp` to `amqps` and provide
specific user credentials.