Add config migration to add default message broker configuration
The migration simply adds the new broker related fields to each profile
using the default value, unless the profile already defines it. The
migration version is upped, but the last backwards compatible version is
kept the same. This is because if the configuration is used with version
3 of the config file, the new keys are simply removed as the config file
is parsed. When the code is updated again, the new keys are added again
using the defaults.